% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{aves_tk4tel}
\alias{aves_tk4tel}
\title{Aves gridded data of Bavaria}
\format{
A \code{data.frame} with 4623 observations and 5 variables.
}
\source{
This data has been digitized from the Bavarian Atlas of Birds.
}
\usage{
data(aves_tk4tel)
}
\description{
Aves gridded data of Bavaria, Germany
}
\details{
Aves gridded data of Bavaria, Germany
}
