% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavDC-package.R
\docType{package}
\name{bavDC package}
\alias{bavDC package}
\alias{bavDCpackage}
\title{bavDC: Bavarian Data Cube}
\description{
R package to access various data sources for Bavaria, i.e. elevation, climate and land-use data.
}
\details{
bavDC: Bavarian Data Cube
}
\author{
RS-eco
}
\keyword{package}
