% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavDC-package.R
\docType{data}
\name{dwd_annual_ts_bav}
\alias{dwd_annual_ts_bav}
\title{DWD annual time series data of Bavaria}
\format{
A \code{sf} object with 70 observations and 11 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://opendata.dwd.de/climate_environment/CDC/regional_averages_DE/annual/}}
}
\usage{
data(dwd_annual_ts_bav)
}
\description{
DWD annual time series climate data of Bavaria
}
\details{
Temperature, Frost days, precipitation data from the Deutscher Wetter Dienst 
summarised to annual data for Bavaria.
}
