% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavDC-package.R
\docType{data}
\name{e-obs_bioclim_1969_1999_bav}
\alias{e-obs_bioclim_1969_1999_bav}
\title{E-OBS bioclimatic data for Bavaria at 0.1 degree resolution from 1969 to 1999}
\format{
A \code{data.frame} with 870 observations and 21 variables.
}
\usage{
data(`e-obs_bioclim_1969_1999_bav`)
}
\description{
E-OBS gridded bioclimatic data for Bavaria at 0.1 degree resolution from 1969 to 1999
}
\details{
E-OBS gridded bioclimatic data for Bavaria at 0.1 degree resolution from 1969 to 1999.
}
