% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{e-obs_bioclim_1979_2009_bav}
\alias{e-obs_bioclim_1979_2009_bav}
\title{E-OBS bioclimatic data for Bavaria at 0.1 degree resolution from 1979 to 2009}
\format{
A \code{data.frame} with 870 observations and 21 variables.
}
\usage{
data(`e-obs_bioclim_1979_2009_bav`)
}
\description{
E-OBS gridded bioclimatic data for Bavaria at 0.1 degree resolution from 1979 to 2009
}
\details{
E-OBS gridded bioclimatic data for Bavaria at 0.1 degree resolution from 1979 to 2009.
}
