% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{e-obs_elev_ens_0.1deg_bav}
\alias{e-obs_elev_ens_0.1deg_bav}
\title{E-OBS gridded elevation data for Bavaria at 0.1 degree resolution}
\format{
A \code{data.frame} with 870 observations and 3 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://cds.climate.copernicus.eu/cdsapp#!/dataset/insitu-gridded-observations-europe?tab=form}}
}
\usage{
data(`e-obs_elev_ens_0.1deg_bav`)
}
\description{
E-OBS gridded elevation data for Bavaria at 0.1 degree resolution
}
\details{
E-OBS gridded elevation data for Bavaria at 0.1 degree resolution.
}
