% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{e-obs_rr_ens_mean_0.1deg_bav_yearmon}
\alias{e-obs_rr_ens_mean_0.1deg_bav_yearmon}
\title{E-OBS gridded precipitation amount ensemble mean data for Bavaria at 0.1 degree resolution}
\format{
A \code{data.frame} with 870 observations and 854 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://cds.climate.copernicus.eu/cdsapp#!/dataset/insitu-gridded-observations-europe?tab=form}}
}
\usage{
data(`e-obs_rr_ens_mean_0.1deg_bav_yearmon`)
}
\description{
E-OBS gridded precipitation amount ensemble mean data for Bavaria at 0.1 degree resolution
}
\details{
E-OBS daily gridded precipitation amount ensemble mean of observations 
for Bavaria from 1950 to present at 0.1 degree resolution.
Total daily amount of rain, snow and hail measured as the height of the equivalent liquid water 
in a square meter The data sources for the precipitation are rain gauge data which 
do not have a uniform way of defining the 24-hour period over which precipitation 
measurements are made. Therefore, there is no uniform time period 
(for instance, 06 UTC previous day to 06 UTC today) which could be attached 
to the daily precipitation.
}
