% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavDC-package.R
\docType{data}
\name{e-obs_tx_ens_mean_0.1deg_bav_yearmon}
\alias{e-obs_tx_ens_mean_0.1deg_bav_yearmon}
\title{E-OBS gridded maximum temperature ensemble mean data for Bavaria at 0.1 degree resolution}
\format{
A \code{data.frame} with 870 observations and 854 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://cds.climate.copernicus.eu/cdsapp#!/dataset/insitu-gridded-observations-europe?tab=form}}
}
\usage{
data(`e-obs_tx_ens_mean_0.1deg_bav_yearmon`)
}
\description{
E-OBS gridded maximum temperature ensemble mean data for Bavaria at 0.1 degree resolution
}
\details{
E-OBS daily gridded maximum temperature ensemble mean of observations 
for Bavaria from 1950 to present at 0.1 degree resolution.
Maximum air temperature measured near the surface, usually at height of 2 meters.
}
