% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{ebcc_bav}
\alias{ebcc_bav}
\title{European Bird Census Council data of bavaria}
\format{
A \code{data.frame} with 45 observations and 497 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{http://www.gbif.org}}. 
More information can be found here: \itemize{\item \url{https://www.ebcc.info/what-we-do/ebba2/ebcc-atlas-of-european-breeding-birds/}}.
}
\usage{
data(ebcc_bav)
}
\description{
EBCC bird data of Bavaria, Germany
}
\details{
Euoprean Breeding Bird Atlas data of Bavaria.
}
