% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{eu_dem_bav_500m}
\alias{eu_dem_bav_500m}
\title{altitudinal data of bavaria}
\format{
A \code{data.frame} with 285106 observations and 7 variables.
}
\source{
This data has been obtained from: 
\itemize{\item \url{https://ec.europa.eu/eurostat/de/web/gisco/geodata/reference-data/elevation/eu-dem/eu-dem-laea}}
}
\usage{
data(eu_dem_bav_500m)
}
\description{
EU-DEM elevation data of Bavaria, Germany
}
\details{
Elevation data of Bavaria, Germany derived from EU-DEM data and 
aggregated to 500 m resolution.
}
