% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{hillshade_bav_500m}
\alias{hillshade_bav_500m}
\title{hillshade data of bavaria}
\format{
A \code{data.frame} with 285106 observations and 7 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://ec.europa.eu/eurostat/de/web/gisco/geodata/reference-data/elevation/eu-dem/hillshade}}
}
\usage{
data(hillshade_bav_500m)
}
\description{
EU-DEM hillshade data of Bavaria, Germany
}
\details{
Hillshade data of Bavaria, Germany derived from EU-DEM data and 
aggregated to 500 m resolution.
}
