% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{inat_bav_1955_2021}
\alias{inat_bav_1955_2021}
\title{iNaturalist data of bavaria}
\format{
A \code{data.frame} object with 234928 observations and 9 variables.
}
\source{
This data has been obtained through the rinat package from: 
\itemize{\item \url{https://www.inaturalist.org/}}
}
\usage{
data(inat_bav_1955_2021)
}
\description{
iNatrualist data of Bavaria, Germany for 1955 - 2021
}
\details{
iNatrualist occurrence data of Bavaria, Germany for 1955 - 2021.
}
