% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{ng_abspziele}
\alias{ng_abspziele}
\title{Naturraum-Untereinheiten Ziele of Bavaria}
\format{
A \code{data.frame} with 834 observations and 4 variables.
}
\source{
This data has been obtained from: 
\itemize{\item \url{https://www.lfu.bayern.de/natur/naturraeume/index.htm}}
}
\usage{
data(ng_abspziele)
}
\description{
Naturraum-Untereinheiten Ziele of Bavaria, Germany
}
\details{
Naturraum-Untereinheiten of Bavaria, Germany
}
