% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavDC-package.R
\docType{data}
\name{ng_geo}
\alias{ng_geo}
\title{Biogeographic regions of Bavaria}
\format{
A \code{data.frame} with 2 observations and 3 variables.
}
\source{
This data has been obtained from: 
\itemize{\item \url{https://www.lfu.bayern.de/natur/naturraeume/index.htm}}
}
\usage{
data(ng_geo)
}
\description{
Biogeographic regions of Bavaria, Germany
}
\details{
Biogeographic regions of Bavaria, Germany
}
