% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavDC-package.R
\docType{data}
\name{ng_gross}
\alias{ng_gross}
\title{Grosslandschaften of Bavaria}
\format{
A \code{data.frame} with 5 observations and 4 variables.
}
\source{
This data has been obtained from: 
\itemize{\item \url{https://www.lfu.bayern.de/natur/naturraeume/index.htm}}
}
\usage{
data(ng_gross)
}
\description{
Grosslandschaften of Bavaria, Germany
}
\details{
Grosslandschaften of Bavaria, Germany
}
