% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavDC-package.R
\docType{data}
\name{ng_meynen_schmit}
\alias{ng_meynen_schmit}
\title{Naturraum-Einheiten of Bavaria}
\format{
A \code{data.frame} with 96 observations and 9 variables.
}
\source{
This data has been obtained from: 
\itemize{\item \url{https://www.lfu.bayern.de/natur/naturraeume/index.htm}}
}
\usage{
data(ng_meynen_schmit)
}
\description{
Naturraum-Einheiten of Bavaria, Germany
}
\details{
Naturraum-Einheiten of Bavaria, Germany
}
