% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavDC-package.R
\docType{data}
\name{ng_ssymank}
\alias{ng_ssymank}
\title{Naturraum-Haupteinheiten of Bavaria}
\format{
A \code{data.frame} with 18 observations and 7 variables.
}
\source{
This data has been obtained from: 
\itemize{\item \url{https://www.lfu.bayern.de/natur/naturraeume/index.htm}}
}
\usage{
data(ng_ssymank)
}
\description{
Naturraum-Haupteinheiten of Bavaria, Germany
}
\details{
Naturraum-Haupteinheiten of Bavaria, Germany
}
