% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{ng_unter_absp}
\alias{ng_unter_absp}
\title{Naturraum-Untereinheiten of Bavaria}
\format{
A \code{data.frame} with 354 observations and 10 variables.
}
\source{
This data has been obtained from: 
\itemize{\item \url{https://www.lfu.bayern.de/natur/naturraeume/index.htm}}
}
\usage{
data(ng_unter_absp)
}
\description{
Naturraum-Untereinheiten of Bavaria, Germany
}
\details{
Naturraum-Untereinheiten of Bavaria, Germany
}
