% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavDC-package.R
\docType{data}
\name{odonata_tk4tel}
\alias{odonata_tk4tel}
\title{Odonata gridded data of Bavaria}
\format{
A \code{data.frame} with 23119 observations and 5 variables.
}
\source{
This data has been digitized from: 
\itemize{\item \url{Kuhn, K. & Burbach, K. (Eds.).(2008). 
Libellen in Bayern. Verlag Eugen Ulmer, Stuttgart.}}
}
\usage{
data(odonata_tk4tel)
}
\description{
Odonata gridded data of Bavaria, Germany
}
\details{
Odonata gridded data of Bavaria, Germany
}
