% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{pnv_bav}
\alias{pnv_bav}
\title{Potential natural vegetation of Bavaria}
\format{
A \code{data.frame} with 212 observations and 9 variables.
}
\source{
This data has been obtained from: 
\itemize{\item \url{https://www.lfu.bayern.de/natur/potentielle_natuerliche_vegetation/index.htm}}
}
\usage{
data(pnv_bav)
}
\description{
Potential natural vegetation of Bavaria, Germany
}
\details{
Potential natural vegetation of Bavaria, Germany
}
