% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bavDC-package.R
\docType{data}
\name{tk4tel_db}
\alias{tk4tel_db}
\title{tk4tel database data of bavaria}
\format{
A \code{data.frame} with 2285 observations and 11 variables.
}
\usage{
data(tk4tel_db)
}
\description{
TK4tel database of Bavaria, Germany
}
\details{
TK4tel database coordinate data of Bavaria, Germany derived from the LFU Biodiversity database.
}
