% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc-package.R
\docType{data}
\name{wfde5_tair_bav}
\alias{wfde5_tair_bav}
\title{WFDE5 reanalysis yearmon air temperature data of Europe}
\format{
A \code{data.frame} with 47 observations and 494 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://cds.climate.copernicus.eu/cdsapp#!/dataset/10.24381/cds.20d54e34?tab=form}}
}
\usage{
data(wfde5_tair_bav)
}
\description{
WFDE5 reanalysis yearmon air temperature data of Europe
}
\details{
WFDE5 reanalysis yearmon near-surface air temperature in K, 
the temperature of air at 2 metres above the surface of land, sea or inland waters, data of Europe.
The data.frame contains air temperature data from January 1979 - December 2019 at a 0.5° grid
}
